backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-dotenv==1.0.0
sqlalchemy==2.0.23
asyncpg==0.29.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.107.0
anthropic==0.25.0
langgraph==0.0.41
langchain==0.1.9
pgvector==0.2.4
psycopg2-binary==2.9.9
cryptography==41.0.7
pyjwt==2.8.1
